backend/api/requirements.txt
python-dotenv>=0.14.0
Flask>=3.0.0
Flask-HTTPAuth>=4.8.0
pandas==2.1.4
waitress>=3.0.0
paho-mqtt==2.0.0
